y/requirements.txt
gradio>=4.24.0
pandas==2.2.0
numpy==1.26.3
reportlab==4.1.0
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
google-genai>=0.1.0
websockets>=13.0.0,<15.1.0
chromadb==0.4.24
sentence-transformers==2.5.1 